+++ pkg/service/apis/plugins/plugins.go
@@ -2,9 +2,6 @@ package plugins
 
 import (
 	"github.com/emicklei/go-restful/v3"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"kubegems.io/kubegems/pkg/utils/httputil/response"
 	"kubegems.io/kubegems/pkg/utils/kube"
 	"kubegems.io/kubegems/pkg/utils/route"
@@ -39,22 +36,8 @@ func (p *PluginsAPI) List(req *restful.Request, resp *restful.Response) {
 	ret := []PluginsStatus{}
 
 	// kubegems
-	svc := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kubegems-models-store",
-			Namespace: "kubegems",
-		},
-	}
-	if err := p.cli.Get(req.Request.Context(), client.ObjectKeyFromObject(svc), svc); err != nil {
-		if errors.IsNotFound(err) {
-			ret = append(ret, PluginsStatus{Name: "kubegems-models", Enabled: false})
-		} else {
-			response.Error(resp, err)
-			return
-		}
-	} else {
-		ret = append(ret, PluginsStatus{Name: "kubegems-models", Enabled: true})
-	}
+	// TODO: check if plugin is enabled
+	ret = append(ret, PluginsStatus{Name: "kubegems-models", Enabled: true})
 	response.OK(resp, ret)
 }
